moved from json to convert pulled SchemaNode from the script into new library made use of new codegen features for consistency (generate test libraries) deleted root_finder (function in utils.dart and autogenerated) deleted test_runner (codegen runner replaces it and hop ties it in)
