+++ test/runner.dart
@@ -1,7 +1,7 @@
 import 'package:unittest/unittest.dart';
 import 'package:unittest/vm_config.dart';
-import "test_validation.dart" as test_validation;
-import "test_invalid_schemas.dart" as test_invalid_schemas;
+import 'test_invalid_schemas.dart' as test_invalid_schemas;
+import 'test_validation.dart' as test_validation;
 
 void testCore(Configuration config) {
   unittestConfiguration = config;
@@ -12,3 +12,4 @@ main() {
   test_invalid_schemas.main();
   test_validation.main();
 }
+
