+++ test/test_invalid_schemas.dart
@@ -1,31 +1,33 @@
-import "dart:io";
-import "root_finder.dart";
-import "dart:json" as JSON;
+library test_invalid_schemas;
+
+import 'dart:io';
+import 'dart:convert' as convert;
+import 'package:path/path.dart' as path;
+import 'package:json_schema/json_schema.dart';
 import 'package:unittest/unittest.dart';
-import "package:json_schema/json_schema.dart";
-import "package:path/path.dart" as path;
-import "package:logging/logging.dart";
-import "package:logging_handlers/logging_handlers_shared.dart";
+import 'package:logging/logging.dart';
+// custom <additional imports>
+import 'utils.dart';
+// end <additional imports>
 
-final _logger = new Logger("test");
 
-main() {
+final _logger = new Logger("test_invalid_schemas");
 
-  ////////////////////////////////////////////////////////////////////////
-  // Uncomment to see logging
-  //
-  // Logger.root.onRecord.listen(new PrintHandler());
-  // Logger.root.level = Level.FINE;
-  // logFormatExceptions = true;
+// custom <library test_invalid_schemas>
+// end <library test_invalid_schemas>
 
-  String here = rootFinder('json_schema');
+main() { 
+// custom <main>
+
+  String here = packageRootPath;
   Directory testSuiteFolder = new Directory("${here}/test/invalid_schemas");
 
   testSuiteFolder.listSync().forEach((testEntry) {
     String shortName = path.basename(testEntry.path);
     group("Invalid schema: ${shortName}", () {
       if(testEntry is File) {
-        List tests = JSON.parse((testEntry as File).readAsStringSync());
+        List tests = 
+          convert.JSON.decode((testEntry as File).readAsStringSync());
         tests.forEach((testObject) {
           var schemaData = testObject["schema"];
           var description = testObject["description"];
@@ -48,4 +50,7 @@ main() {
       }
     });
   });
+
+// end <main>
+
 }
