+++ test/utils.dart
@@ -1,0 +1,15 @@
+import 'dart:io';
+import 'package:path/path.dart' as path;
+
+String get packageRootPath {
+  var parts = path.split(path.absolute(new Options().script));
+  int found = parts.lastIndexOf('json_schema');
+  if(found >= 0) {
+    return path.joinAll(parts.getRange(0, found+1));
+  }
+  throw new 
+    StateError("Current directory must be within package 'json_schema'");
+}
+
+main() => print(packageRootPath);
+
